Integrate from development (#19)
* Add lottery concept

* One of the long lasting requests from BXH: reproducable optimizations.**

** only for single thread, only for clusters

In this commit: core and generic are converted to use the Lottery singleton which can make use of different RNG backends. To add another RNG: implement interface and add to the switch in Input.

Please note: if you are NOT specifying RandomNumberGenerator=, it'll just use a Java RNG seeded with a random number gotten from... a Java RNG. It'll provide the sed value on System.out, so if one was to then reproduce this run, one would say (if the seed was -34354647575): RandomNumberGenerator=javarng:seed=-34354647575

* Also break the loop here, looping takes a lot of time.

* Use fma for routines that are expensive for LBFGS.

Will increase precision and on any x86 system younger than 4 years old will improve performance.