XTB as backend (energy only)
Add support for energy evalutation to XTB, i.e., formally make it a CartesianFullBackend. However, due to the lack of a gradient implementation, this functionality will not be exposed but only used from operators that may need energies.

While there, adjust the precision for the XTB macrobenchmark to better match the opt normal setting.